requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
apptools==5.1.0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
configobj==5.0.6
constantly==15.1.0
cryptography==35.0.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
envisage==6.0.1
flatbuffers==2.0
fsspec==2021.7.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.33.1
google-auth-oauthlib==0.4.4
graphviz==0.17
grpcio==1.39.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.3.0
incremental==21.3.0
ipdb==0.13.9
ipyevents==2.0.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jax==0.2.18
jaxlib==0.1.69
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==1.5.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mayavi==4.7.3
mistune==0.8.4
multidict==5.1.0
munch==2.5.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
pyface==7.3.0
Pygments==2.9.0
pymango==0.1.1
pymongo==4.0.1
pyparsing==2.4.7
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.5.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacred==0.8.2
scikit-learn==0.24.2
scipy==1.7.0
seml @ git+https://github.com/TUM-DAML/seml@a4740c6a7499fe62e9775e5d10216008774af85c
Send2Trash==1.7.1
six==1.16.0
smmap==5.0.0
statsmodels==0.12.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.0
torch-tb-profiler==0.2.1
torchinfo==1.5.2
torchmetrics==0.4.1
torchsummary==1.5.1
torchviz==0.0.2
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
traits==6.3.1
traitsui==7.2.1
tueplots==0.0.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vtk==9.0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.13.3
wslink==1.1.0
yarl==1.6.3
zipp==3.6.0
zope.interface==5.4.0
